Refactored package name, changes accordingly
