+++ src/View/BaseFrame.java
@@ -1,6 +1,6 @@
 package View;
 
-import db_objs.User;
+import Database.User;
 
 import javax.swing.*;
